requirements.txt
numpy
matplotlib
jupyterlab
python-dotenv
